clean up directory structure
