+++ src/vivarium_nih_us_cvd/components/__init__.py
@@ -2,6 +2,6 @@ from .disease import IschemicHeartDiseaseAndHeartFailure, IschemicStroke
 from .effects import InterventionAdherenceEffect
 from .healthcare_utilization import HealthcareUtilization
 from .interventions import LinearScaleUp
-from .observers import PAFObserver
+from .paf_observer import PAFObserver
 from .risks import AdjustedRisk, TruncatedRisk
 from .treatment import Treatment
